refactor: RoleServiceTests - refactor checkGetAll method; SpringBootTest - refactor for transactions; Transaction - remove field card and add String for recipientCard number, added equals and hashCode + toString method add account; TransactionService - refactor int sum to double sum in method create; TransactionServiceImpl - refactor all create method to better reading and productivity; TransactionTests - added recipientCard to validTransaction field. added: TransactionServiceTests - where I write all tests for TransactionService methods

+++ src/main/java/com/pet/project/service/TransactionService.java
@@ -5,7 +5,7 @@ import com.pet.project.model.entity.Transaction;
 import java.util.List;
 
 public interface TransactionService {
-    Transaction create(Transaction transaction, int sum);
+    Transaction create(Transaction transaction, double sum);
     void delete(long id);
     Transaction update(Transaction transaction);
     Transaction readById(long id);
